requirements.txt
django
djangorestframework
django-stubs
djangorestframework-simplejwt
python-dotenv